Use uint8_t instead of ambiguous int for colour

+++ model/team.c
@@ -3,7 +3,7 @@
 #include <string.h>
 #include <wchar.h>
 
-Team * createTeam(char * name, int colour, Piece * pieces, uint8_t piece_count) {
+Team * createTeam(char * name, uint8_t colour, Piece * pieces, uint8_t piece_count) {
     Team * out = malloc(sizeof(Team));
     strcpy(out->name, name);
     out->colour = colour;
